+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased Changes
 
  * Added option to display table-of-contents on namespace pages
+ * Improve formatting of large objects when used as a card body
 
 ## v0.6.2